Replace `buildRustPackage` with `mkDerivation`
It looks like a lot of assumptions that don't apply to my case are build
into `buildRustPackage`, causing it to get in the way a lot. Building on
the lower layer that `buildRustPackage` itself builds on might be a
better solution.

At the very least, I've gotten farther than before: The build is
running, even though it does nothing.